Organized folders to match Atomic Design
