+++ src/components/atoms/Button.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { withRouter } from 'react-router-dom';
 // import '../stylesheets/components/Header.css'; 
-import '../stylesheets/main.css'; // for dev
+import '../../stylesheets/main.css'; // for dev
 
 class Button extends Component {
   constructor(props){
