+++ src/components/atoms/Input.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import '../stylesheets/main.css'; // for dev
+import '../../stylesheets/main.css';  // for dev
 
 class Input extends Component {
   onChange = (e) => {
