+++ src/components/atoms/Loader.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import '../stylesheets/main.css'; 
+import '../../stylesheets/main.css'; 
 import loader from "../images/loader.gif";
