+++ src/components/molecules/Header.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { withRouter } from 'react-router-dom';
-import '../stylesheets/main.css'; // for dev
-import Button from './Button.js';
+import '../../stylesheets/main.css';  // for dev
+import Button from '../atoms/Button.js';
 
 class Header extends Component{
   render(){
