+++ src/components/molecules/ProjectInfo.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react';
 // import '../stylesheets/components/ProjectInfo.css';
-import '../stylesheets/main.css'; // for dev
-import Button from './Button.js';
+import '../../stylesheets/main.css'; // for dev
+import Button from '../atoms/Button.js';
 import ProjectCard from './ProjectCard';
+import Loader from "../atoms/Loader";
 
 class ProjectInfo extends Component {
     constructor(props) {
@@ -63,7 +64,7 @@ class ProjectInfo extends Component {
             const isOwner = (user && owner && user._id === owner._id);
             
             if (!project) {
-                return <h3>Loading...</h3>;
+                return <Loader />
             }
             
             let buttons = null;
