+++ src/components/molecules/UserEdit.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 // import '../stylesheets/components/UserEdit.css';
-import '../stylesheets/main.css'; // for dev
-import Button from './Button';
-import Input from './Input';
+import '../../stylesheets/main.css'; // for dev
+import Button from '../atoms/Button';
+import Input from '../atoms/Input';
 
 
 class UserEditNew extends Component {
