+++ src/components/organisms/About.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import { withRouter } from 'react-router-dom';
-import '../stylesheets/main.css'; 
-import Button from './Button.js';
-import target from "../images/target.svg";
-import startup from "../images/startup.svg";
+import '../../stylesheets/main.css'; 
+import Button from '../atoms/Button.js';
+import target from "../../images/target.svg";
+import startup from "../../images/startup.svg";
 
 
 class About extends Component{
